Add files via upload
Hi All,
I am having problems trying to get these files integrated into the rails app. Unfortunately, I still struggle with Rails so I uploaded these separately.
These files are an HTML display, a javascript page that is src'ed into the html page, and a CSS page.
The most important part of these files is the .js file. This contains the Javascript logic that should be able to be implemented into the Rails app.
The coloring for the page is just a test to see if anyone like the color scheme. I went with orange and crimson red, I think the colors go well together. Maybe others like it as well? I was thinking a stripe design to resemble the tail of a Tanooki. If you like these colors,  the red is rgb(201,0,22) and the orange is simply orange.
Good luck this weekend, I am sorry I cannot make it but look forward to attending the next meetings, hopefully I won't have to miss any other ones. It would be helpful for me to see this integrated into Rails